Squashed commit of the following:
commit bb546df5983235ef70d70e38b292c613d4d67b70
Merge: 42d345f ae5b906
Author: Prateek Kumar <prateek@prateekkumar.in>
Date:   Wed Apr 24 23:33:58 2019 +0530

    Merge branch 'file_type' of github.com:vbsinha/pdp-judge into file_type

commit ae5b90676d759131bac1b9051a387ea1977eae89
Merge: 63b5b74 ce27d2b
Author: vishwakftw <cs15btech11043@iith.ac.in>
Date:   Wed Apr 24 23:33:29 2019 +0530

    Merge branch 'file_type' of github.com:vbsinha/pdp-judge into file_type

commit 63b5b7477b85f04ddf95339aa3ec3c34a1d3cb95
Author: vishwakftw <cs15btech11043@iith.ac.in>
Date:   Wed Apr 24 23:32:59 2019 +0530

    Orphaning contest not allowed

commit 42d345fafea0bf47ea52db7db88b4fa4512d24d6
Author: Prateek Kumar <prateek@prateekkumar.in>
Date:   Wed Apr 24 23:32:50 2019 +0530

    settings: allowed hosts make it *


commit ce27d2bbf2b733f987c4e57cfa13c9c6c934826b
Author: Prateek Kumar <prateek@prateekkumar.in>
Date:   Wed Apr 24 22:51:29 2019 +0530

    Display delete poster button for posters only


commit 826e1cf8a0d2d375a3e01f7d7671322d9237044b
Author: vishwakftw <cs15btech11043@iith.ac.in>
Date:   Wed Apr 24 22:49:08 2019 +0530

    Enable file_type checking

commit f70c55c4f4731cdace86ecaccd4f5ad5ac298c39
Merge: 9f13922 9d04179
Author: vishwakftw <cs15btech11043@iith.ac.in>
Date:   Wed Apr 24 22:30:09 2019 +0530

    Merge branch 'file_type' of github.com:vbsinha/pdp-judge into file_type

commit 9f13922916ccd69e3e12dcc0b7645aa3609ca421
Author: vishwakftw <cs15btech11043@iith.ac.in>
Date:   Wed Apr 24 22:29:54 2019 +0530

    Compute score in pylint

commit 9d0417954c2ea9f839d293f06b686359e94cc3fe
Author: Prateek Kumar <prateek@prateekkumar.in>
Date:   Wed Apr 24 22:29:24 2019 +0530

    Add file type of submission

+++ judge/handler.py
@@ -207,11 +207,13 @@ def process_solution(problem_id: str, participant: str, file_type,
     problem is the 'code' (pk) of the problem. participant is email(pk) of the participant
     """
     try:
-        file_type = '.py'  # TODO file_type
         problem = models.Problem.objects.get(pk=problem_id)
         participant = models.Person.objects.get(email=participant)
         s = problem.submission_set.create(participant=participant, file_type=file_type,
                                           submission_file=submission_file, timestamp=timestamp)
+        if file_type not in problem.file_format.split(','):
+            return (False, 'Accepted file types: \"{}\"'
+                           .format(', '.join(problem.file_format.split(','))))
         s.save()
     except Exception as e:
         traceback.print_exc()
@@ -321,8 +323,11 @@ def delete_personcontest(person: str, contest: str) -> Tuple[bool, Optional[str]
     try:
         p = models.Person.objects.get(email=person)
         c = models.Contest.objects.get(pk=contest)
-        models.ContestPerson.objects.filter(person=p, contest=c).delete()
-        return (True, None)
+        if models.ContestPerson.objects.filter(contest=c).count() > 1:
+            models.ContestPerson.objects.filter(person=p, contest=c).delete()
+            return (True, None)
+        else:
+            return (False, 'This contest cannot be orphaned!')
     except Exception as e:
         traceback.print_exc()
         return (False, e.__str__())
